# Files regex processing boost