--- a/src/NuGet.Core/NuGet.Common/PathUtil/PathResolver.cs
+++ b/src/NuGet.Core/NuGet.Common/PathUtil/PathResolver.cs
@@ -23,13 +23,13 @@ namespace NuGet.Common
             Func<T, string> getPath,
             IEnumerable<string> wildcards)
         {
-            IEnumerable<Regex> filters = wildcards.Select(WildcardToRegex);
+            IEnumerable<Regex> filters = wildcards.Select(WildcardToRegex).ToList();
 
-            return source.Where(item =>
+            return source.AsParallel().Where(item =>
             {
                 string path = getPath(item);
                 return filters.Any(f => f.IsMatch(path));
-            });
+            }).ToList();
         }
 
         /// <summary>
@@ -80,6 +80,7 @@ namespace NuGet.Common
                 pattern = pattern
                     .Replace(@"\.\*\*", @"\.[^/.]*") // .** should not match on ../file or ./file but will match .file
                     .Replace(@"\*\*/", "(.+/)*") //For recursive wildcards /**/, include the current directory.
+                    .Replace(@"\*\*/", "([^/]+/)*") //For recursive wildcards /**/, include the current directory.
                     .Replace(@"\*\*", ".*") // For recursive wildcards that don't end in a slash e.g. **.txt would be treated as a .txt file at any depth
                     .Replace(@"\*", @"[^/]*(/)?") // For non recursive searches, limit it any character that is not a directory separator
                     .Replace(@"\?", "."); // ? translates to a single any character
@@ -88,12 +89,12 @@ namespace NuGet.Common
             {
                 // regex wildcard adjustments for Windows-style file systems
                 pattern = pattern
-                    .Replace("/", @"\\") // On Windows, / is treated the same as \.
-                    .Replace(@"\.\*\*", @"\.[^\\.]*") // .** should not match on ../file or ./file but will match .file
-                    .Replace(@"\*\*\\", @"(.+\\)*") //For recursive wildcards \**\, include the current directory.
-                    .Replace(@"\*\*", ".*") // For recursive wildcards that don't end in a slash e.g. **.txt would be treated as a .txt file at any depth
-                    .Replace(@"\*", @"[^\\]*(\\)?") // For non recursive searches, limit it any character that is not a directory separator
-                    .Replace(@"\?", "."); // ? translates to a single any character
+                    .Replace("/", @"\\")                 // On Windows, / is treated the same as \.
+                    .Replace(@"\.\*\*", @"\.[^\\.]*")    // .** should not match on ../file or ./file but will match .file
+                    .Replace(@"\*\*\\", @"([^\\]+\\)*?") //For recursive wildcards \**\, include the current directory.
+                    .Replace(@"\*\*", ".*")              // For recursive wildcards that don't end in a slash e.g. **.txt would be treated as a .txt file at any depth
+                    .Replace(@"\*", @"[^\\]*(\\)?")      // For non recursive searches, limit it any character that is not a directory separator
+                    .Replace(@"\?", ".");                // ? translates to a single any character
             }
 
             return new Regex('^' + pattern + '$', RegexOptions.IgnoreCase | RegexOptions.ExplicitCapture | RegexOptions.CultureInvariant);
@@ -146,7 +147,7 @@ namespace NuGet.Common
 
             // Starting from the base path, enumerate over all files and match it using the wildcard expression provided by the user.
             // Note: We use Directory.GetFiles() instead of Directory.EnumerateFiles() here to support Mono
-            IEnumerable<SearchPathResult> matchedFiles = from file in Directory.GetFiles(normalizedBasePath, "*.*", searchOption)
+            IEnumerable<SearchPathResult> matchedFiles = from file in Directory.GetFiles(normalizedBasePath, "*.*", searchOption).AsParallel()
                                                          where searchRegex.IsMatch(file)
                                                          select new SearchPathResult(file, isFile: true);