--- a/test/NuGet.Core.Tests/NuGet.Common.Test/PathResolverTests.cs
+++ b/test/NuGet.Core.Tests/NuGet.Common.Test/PathResolverTests.cs
@@ -146,7 +146,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { "a{0}d", "a{0}b{0}d", "a{0}b{0}c{0}d" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -157,7 +160,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { ".{0}.c", "a{0}.c", "a{0}{0}.c", "b.c", "a{0}b{0}bc.c", "a{0}b{0}.c" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -168,7 +174,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { ".{0}.c", "a{0}.c", "a{0}{0}.c", "b.c", "a{0}b{0}bc.c", "a{0}b{0}.c" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -179,7 +188,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { ".{0}c", "a{0}c" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -190,7 +202,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { "a{0}d", "a{0}b{0}d", "a{0}b{0}c{0}d" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]