--- a/test/NuGet.Core.Tests/NuGet.Common.Test/PathResolverTests.cs
+++ b/test/NuGet.Core.Tests/NuGet.Common.Test/PathResolverTests.cs
@@ -135,7 +135,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { ".c" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -216,7 +219,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             IEnumerable<string> expectedResults = GetPlatformSpecificPaths(new[] { "a{0}b{0}c{0}d" });
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]
@@ -227,7 +233,10 @@ namespace NuGet.Common.Test
             IEnumerable<string> actualResults = PathResolver.GetMatches(sources, source => source, wildcards);
             string[] expectedResults = new[] { "abc", "adc" };
 
-            Assert.Equal(expectedResults, actualResults);
+            IEnumerable<string> orderedActualResults = actualResults.OrderBy(path => path);
+            IEnumerable<string> orderedExpectedResults = expectedResults.OrderBy(path => path);
+
+            Assert.Equal(orderedExpectedResults, orderedActualResults);
         }
 
         [Fact]