Changed recursive wildcards regex pattern to faster form
Added exclude pattern TPL provessing
Fixed linq lazy regex objects spawn

<!-- DO NOT MODIFY OR DELETE THIS TEMPLATE. IT IS USED IN AUTOMATION. -->
## Bug

https://github.com/NuGet/Home/issues/5016

https://github.com/NuGet/Home/issues/5706

Fixes:

Fixed include and exclude files performance

Regression? Last working version:

## Description

Changed regex pattern for recursive wildcards to faster one. 
Lazy quantifier in this case much faster. (.+\\)* -> ([^\\]+\\)*?

Added parallel processing in source files filtering 
https://docs.microsoft.com/en-us/dotnet/api/system.linq.parallelenumerable.asparallel?view=net-5.0

Fixed linq lazy regex objects spawn. 
Lazy filters enumeration in GetMatches method was called on every file entry. ([Source file](https://github.com/NuGet/NuGet.Client/blob/97311b5cb48d38402b49fd41a6384b11db7dd453/src/NuGet.Core/NuGet.Common/PathUtil/PathResolver.cs))

## PR Checklist

- [x ] PR has a meaningful title
- [x ] PR has a linked issue.
- [x ] Described changes

- **Tests**
  - [ ] Automated tests added
  - **OR**
  <!-- Describe why you haven't added automation. -->
  - [x ] Test exception

   [PathResolverTests ](https://github.com/NuGet/NuGet.Client/blob/97311b5cb48d38402b49fd41a6384b11db7dd453/test/NuGet.Core.Tests/NuGet.Common.Test/PathResolverTests.cs) already contains all needed tests.

  - **OR**
  - [ ] N/A <!-- Infrastructure, documentation etc. -->

- **Documentation**
  <!-- Please link the PR/issue if appropriate -->
  - [x ] Documentation PR or issue filled
  - **OR**
  - [ ] N/A