@BlackGad 2021-03-14
Here is some test measurements on 6k files pack
-----------

### Initial version

Exclude pattern | GetMatches time | Matches
 ------------- |-------------| -----
**\\*.pdb                             | 00:00:29.8375342 | 0    
**\\*.obj                             | 00:00:28.1902447 | 0    
**\\*.ilk                             | 00:00:27.8622501 | 0    
Sources\\node_modules\\typescript\\**\\* | 00:00:00.0015847 | 169  
Sources\\node_modules\\@types\\**\\*     | 00:00:00.0008627 | 37   
default excludes                     | 00:00:00.0376492 | 0    

### AsParallel only optimization

Exclude pattern | GetMatches time | Matches
 ------------- |-------------| -----
**\\*.pdb                             | 00:00:06.4921811 | 0
**\\*.obj                             | 00:00:06.0926885 | 0
**\\*.ilk                             | 00:00:06.4990284 | 0
Sources\\node_modules\\typescript\\**\\* | 00:00:00.0011430 | 169
Sources\\node_modules\\@types\\**\\*     | 00:00:00.0007241 | 37
default excludes                     | 00:00:00.0101420 | 0

### AsParallel + Regex lazy quantifier

Exclude pattern | GetMatches time | Matches
 ------------- |-------------| -----
**\\*.pdb                             | 00:00:00.0618149 | 0
**\\*.obj                             | 00:00:00.0421215 | 0
**\\*.ilk                             | 00:00:00.0433455 | 0
Sources\\node_modules\\typescript\\**\\* | 00:00:00.0018229 | 169
Sources\\node_modules\\@types\\**\\*     | 00:00:00.0007414 | 37
default excludes                     | 00:00:00.0154888 | 0