@BlackGad 2021-03-15
https://github.com/NuGet/Home/issues/5706 another related defect. Must be fixed as well because of used regex for home directory