@BlackGad 2021-03-16
Added some fixes to existing unit tests because AsParallel linq method not guaranty result order. 
All operations related to PathResolver methods are order independent so I think we can leave result operation as is.