@erdembayar 2021-03-23
@BlackGad 
I just compared `nuget.exe` from your branch with official 5.8 nuget.exe. So far I don't see actual perf improvement it's very minimal

![image](https://user-images.githubusercontent.com/8766776/112228841-28082880-8bef-11eb-8069-d6587820ac26.png)

Here is official `nuget5.8.exe`:
![image](https://user-images.githubusercontent.com/8766776/112229193-c0061200-8bef-11eb-8d8a-3e016fe52f98.png)

Here is `nuget.exe` from your branch:
![image](https://user-images.githubusercontent.com/8766776/112229207-c8f6e380-8bef-11eb-8606-d4e2945ac9d3.png)

For the sake of completeness I measured other 2 `nuget.exe` which mentioned in https://github.com/NuGet/Home/issues/5016, it clearly shows there was regression in `nuget4.1.0.exe` but it looks addressed (please looked at `nuget5.8.exe` result above):

`nuget.3.5.0.exe`
![image](https://user-images.githubusercontent.com/8766776/112229416-2ab74d80-8bf0-11eb-924e-e4dc16da9d04.png)

`nuget.4.1.0.exe`
![image](https://user-images.githubusercontent.com/8766776/112229436-3440b580-8bf0-11eb-80b4-047b6c9b8a94.png)