@BlackGad 2021-03-24
@erdembayar what did exactly you test?
Issue here when you trying to pack with excludes a lot of files and actual nuspec located deep in file system.

Example and benchmarks:

Current directory: "C:\\"

### NuGet Version: 5.8.1.7021
PS command line
```powershell
Measure-Command -Expression {NuGet pack "C:\1\Test.nuspec" -OutputDirectory c:\temp\ -Version 1.0.0}
```
Result
```
Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 6
Milliseconds      : 54
Ticks             : 60540642
TotalDays         : 7.00701875E-05
TotalHours        : 0.0016816845
TotalMinutes      : 0.10090107
TotalSeconds      : 6.0540642
TotalMilliseconds : 6054.0642
```
PS command line
```powershell
Measure-Command -Expression {NuGet pack "C:\Temp\tests\111111\222222\33333\44444\Test.nuspec" -OutputDirectory c:\temp\ -Version 1.0.1}
```
Result
```
Days              : 0
Hours             : 0
Minutes           : 1
Seconds           : 29
Milliseconds      : 521
Ticks             : 895217841
TotalDays         : 0.00103613176041667
TotalHours        : 0.02486716225
TotalMinutes      : 1.492029735
TotalSeconds      : 89.5217841
TotalMilliseconds : 89521.7841
```
### Patched NuGet

PS command line
```powershell
Measure-Command -Expression {C:\Projects\GitHub\BlackGad\NuGet.Client\artifacts\NuGet.CommandLine\16.0\bin\Debug\net472\NuGet.exe pack "C:\1\Test.nuspec" -OutputDirectory c:\temp\ -Version 2.0.0}
```
Result
```
Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 2
Milliseconds      : 972
Ticks             : 29728922
TotalDays         : 3.4408474537037E-05
TotalHours        : 0.000825803388888889
TotalMinutes      : 0.0495482033333333
TotalSeconds      : 2.9728922
TotalMilliseconds : 2972.8922
```
PS command line
```powershell
Measure-Command -Expression {C:\Projects\GitHub\BlackGad\NuGet.Client\artifacts\NuGet.CommandLine\16.0\bin\Debug\net472\NuGet.exe pack "C:\Temp\tests\111111\222222\33333\44444\Test.nuspec" -OutputDirectory c:\temp\ -Version 2.0.1}
```
Result
```
Days              : 0
Hours             : 0
Minutes           : 0
Seconds           : 3
Milliseconds      : 86
Ticks             : 30864824
TotalDays         : 3.57231759259259E-05
TotalHours        : 0.000857356222222222
TotalMinutes      : 0.0514413733333333
TotalSeconds      : 3.0864824
TotalMilliseconds : 3086.4824
```
Comparison table:
CLI | Short path | Long path
----|-----|---
5.8.1| 6.54 sec|1 min 29.521 sec
Patched| 2.297 sec|3.86 sec

This regex fix boost performance even in best cases for current 5.8.1 version also deep nested folder structure is not affect performance so hard.