@BlackGad 2021-03-24
Forget to attach test data

Create 2 files in target folder:

package.json
```json
{
  "name": "tests",
  "version": "1.0.0",
  "description": "",
  "main": "index.js",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1"
  },
  "author": "",
  "license": "ISC",
  "devDependencies": {
    "@types/debug": "^4.1.5",
    "@types/jquery": "^3.3.38",
    "@types/node": "^8.0.14",
    "@types/yargs": "^15.0.4",
    "typescript": "^3.2.2"
  },
  "dependencies": {
    "debug": "^4.1.1",
    "node-red": "^1.2.6",
    "yargs": "^15.3.0"
  }
}
```
Test.nuspec
```xml
<?xml version="1.0"?>
<package>
  <metadata>
    <id>Test</id>
    <version>$version$</version>
    <authors>Volodymyr Shkolka</authors>
    <owners>Volodymyr Shkolka</owners>
    <requireLicenseAcceptance>false</requireLicenseAcceptance>
    <description>Test node-RED npm package pack</description>
  </metadata>
  <files>
    <file src="node_modules\**\*" target="libs\native\node_modules" exclude="**\*.pdb;**\*.obj;**\*.ilk;node_modules\typescript\**\*;node_modules\@types\**\*"/>
  </files>
</package>
```

Then in this folder call 
```
npm install
```
After that you are ready to pack NuGet with nuspec specification
1.