@erdembayar 2021-03-24
@BlackGad 
Thank you for providing good test data. Now I can see good improvement on test data you provided.

![image](https://user-images.githubusercontent.com/8766776/112380634-d7540680-8ca6-11eb-95b3-8a1d1d60deda.png)