@erdembayar 2021-03-24
@zkat  @kartheekp-ms @dtivel 
I can confirm there is perf improvement on test data, looks good for me. But I not sure about security aspect of regex, please give your review, I'm not familiar about regex quirks.