@BlackGad 2021-03-25
Can you share me build details?